requirements.txt
djangorestframework==3.14.0
drf-yasg==1.21.4
Pillow==9.4.0
tensorflow==2.11.0
asgiref==3.6.0
Django==4.1.7
sqlparse==0.4.3
psycopg2-binary==2.9.5
